attempt: tried another setup to grab the rating data
